refactor: 디자인시스템에 Dialog 형식 추가
